+++ packages/safe-ds-lang/src/language/communication/commands.ts
@@ -1,3 +1,4 @@
+export const COMMAND_EXPLORE_TABLE = 'safe-ds.exploreTable';
 export const COMMAND_PRINT_VALUE = 'safe-ds.printValue';
 export const COMMAND_RUN_PIPELINE = 'safe-ds.runPipeline';
 export const COMMAND_SHOW_IMAGE = 'safe-ds.showImage';