+++ packages/safe-ds-lang/src/language/typing/safe-ds-core-types.ts
@@ -26,6 +26,32 @@ export class SafeDsCoreTypes {
         return this.createCoreType(this.builtinClasses.Boolean);
     }
 
+    Cell(wrappedType: Type = this.AnyOrNull): Type {
+        const cell = this.builtinClasses.Cell;
+        const wrappedTypeParameter = getTypeParameters(cell)[0];
+
+        if (!cell || !wrappedTypeParameter) {
+            /* c8 ignore next 2 */
+            return UnknownType;
+        }
+
+        let substitutions = new Map([[wrappedTypeParameter, wrappedType]]);
+        return new ClassType(cell, substitutions, false);
+    }
+
+    Column(elementType: Type): Type {
+        const column = this.builtinClasses.Column;
+        const elementTypeParameter = getTypeParameters(column)[0];
+
+        if (!column || !elementTypeParameter) {
+            /* c8 ignore next 2 */
+            return UnknownType;
+        }
+
+        let substitutions = new Map([[elementTypeParameter, elementType]]);
+        return new ClassType(column, substitutions, false);
+    }
+
     get Float(): Type {
         return this.createCoreType(this.builtinClasses.Float);
     }
@@ -78,6 +104,10 @@ export class SafeDsCoreTypes {
 
     get Number(): Type {
         return this.createCoreType(this.builtinClasses.Number);
+    }
+
+    get Row(): Type {
+        return this.createCoreType(this.builtinClasses.Row);
     }
 
     get String(): Type {