+++ packages/safe-ds-lang/src/language/validation/safe-ds-validator.ts
@@ -160,7 +160,7 @@ import {
     attributeMustHaveTypeHint,
     callReceiverMustBeCallable,
     indexedAccessIndexMustHaveCorrectType,
-    indexedAccessReceiverMustBeListOrMap,
+    indexedAccessReceiverMustHaveCorrectType,
     infixOperationOperandsMustHaveCorrectType,
     listMustNotContainNamedTuples,
     mapMustNotContainNamedTuples,
@@ -189,6 +189,10 @@ import { tagsShouldNotHaveDuplicateEntries } from './builtins/tags.js';
 import { moduleMemberShouldBeUsed } from './other/declarations/moduleMembers.js';
 import { pipelinesMustBePrivate } from './other/declarations/pipelines.js';
 import { thisMustReferToClassInstance } from './other/expressions/this.js';
+import {
+    outputStatementMustHaveValue,
+    outputStatementMustOnlyBeUsedInPipeline,
+} from './other/statements/outputStatements.js';
 
 /**
  * Register custom validation checks.
@@ -294,7 +298,7 @@ export const registerValidationChecks = function (services: SafeDsServices) {
         SdsIndexedAccess: [
             indexedAccessIndexMustBeValid(services),
             indexedAccessIndexMustHaveCorrectType(services),
-            indexedAccessReceiverMustBeListOrMap(services),
+            indexedAccessReceiverMustHaveCorrectType(services),
         ],
         SdsInfixOperation: [
             divisionDivisorMustNotBeZero(services),
@@ -335,6 +339,7 @@ export const registerValidationChecks = function (services: SafeDsServices) {
             namedTypeTypeArgumentListMustNotHavePositionalArgumentsAfterNamedArguments,
             namedTypeTypeArgumentsMustMatchBounds(services),
         ],
+        SdsOutputStatement: [outputStatementMustHaveValue(services), outputStatementMustOnlyBeUsedInPipeline],
         SdsParameter: [
             constantParameterMustHaveConstantDefaultValue(services),
             constantParameterMustHaveTypeThatCanBeEvaluatedToConstant(services),