+++ packages/safe-ds-lang/src/language/validation/types.ts
@@ -99,7 +99,7 @@ export const callReceiverMustBeCallable = (services: SafeDsServices) => {
     };
 };
 
-export const indexedAccessReceiverMustBeListOrMap = (services: SafeDsServices) => {
+export const indexedAccessReceiverMustHaveCorrectType = (services: SafeDsServices) => {
     const typeChecker = services.typing.TypeChecker;
     const typeComputer = services.typing.TypeComputer;
 
@@ -111,7 +111,7 @@ export const indexedAccessReceiverMustBeListOrMap = (services: SafeDsServices) =
 
         const receiverType = typeComputer.computeType(node.receiver);
         if (!typeChecker.canBeAccessedByIndex(receiverType)) {
-            accept('error', `Expected type 'List<T>' or 'Map<K, V>' but got '${receiverType}'.`, {
+            accept('error', `Indexed access is not defined for type '${receiverType}'.`, {
                 node: node.receiver,
                 code: CODE_TYPE_MISMATCH,
             });
@@ -127,10 +127,19 @@ export const indexedAccessIndexMustHaveCorrectType = (services: SafeDsServices) 
 
     return (node: SdsIndexedAccess, accept: ValidationAcceptor): void => {
         const receiverType = typeComputer.computeType(node.receiver);
-        if (typeChecker.isList(receiverType)) {
+        if (typeChecker.isColumn(receiverType) || typeChecker.isList(receiverType)) {
             const indexType = typeComputer.computeType(node.index);
             if (!typeChecker.isSubtypeOf(indexType, coreTypes.Int)) {
                 accept('error', `Expected type '${coreTypes.Int}' but got '${indexType}'.`, {
+                    node,
+                    property: 'index',
+                    code: CODE_TYPE_MISMATCH,
+                });
+            }
+        } else if (typeChecker.isRow(receiverType)) {
+            const indexType = typeComputer.computeType(node.index);
+            if (!typeChecker.isSubtypeOf(indexType, coreTypes.String)) {
+                accept('error', `Expected type '${coreTypes.String}' but got '${indexType}'.`, {
                     node,
                     property: 'index',
                     code: CODE_TYPE_MISMATCH,
@@ -161,17 +170,27 @@ export const infixOperationOperandsMustHaveCorrectType = (services: SafeDsServic
     return (node: SdsInfixOperation, accept: ValidationAcceptor): void => {
         const leftType = typeComputer.computeType(node.leftOperand);
         const rightType = typeComputer.computeType(node.rightOperand);
+        const cellType = coreTypes.Cell();
+
         switch (node.operator) {
             case 'or':
             case 'and':
-                if (node.leftOperand && !typeChecker.isSubtypeOf(leftType, coreTypes.Boolean)) {
-                    accept('error', `Expected type '${coreTypes.Boolean}' but got '${leftType}'.`, {
+                if (
+                    node.leftOperand &&
+                    !typeChecker.isSubtypeOf(leftType, coreTypes.Boolean) &&
+                    !typeChecker.isSubtypeOf(leftType, cellType)
+                ) {
+                    accept('error', `This operator is not defined for type '${leftType}'.`, {
                         node: node.leftOperand,
                         code: CODE_TYPE_MISMATCH,
                     });
                 }
-                if (node.rightOperand && !typeChecker.isSubtypeOf(rightType, coreTypes.Boolean)) {
-                    accept('error', `Expected type '${coreTypes.Boolean}' but got '${rightType}'.`, {
+                if (
+                    node.rightOperand &&
+                    !typeChecker.isSubtypeOf(rightType, coreTypes.Boolean) &&
+                    !typeChecker.isSubtypeOf(rightType, cellType)
+                ) {
+                    accept('error', `This operator is not defined for type '${rightType}'.`, {
                         node: node.rightOperand,
                         code: CODE_TYPE_MISMATCH,
                     });
@@ -199,12 +218,14 @@ export const infixOperationOperandsMustHaveCorrectType = (services: SafeDsServic
             case '-':
             case '*':
             case '/':
+            case '%':
                 if (
                     node.leftOperand &&
                     !typeChecker.isSubtypeOf(leftType, coreTypes.Float) &&
-                    !typeChecker.isSubtypeOf(leftType, coreTypes.Int)
+                    !typeChecker.isSubtypeOf(leftType, coreTypes.Int) &&
+                    !typeChecker.isSubtypeOf(leftType, cellType)
                 ) {
-                    accept('error', `Expected type '${coreTypes.Float}' or '${coreTypes.Int}' but got '${leftType}'.`, {
+                    accept('error', `This operator is not defined for type '${leftType}'.`, {
                         node: node.leftOperand,
                         code: CODE_TYPE_MISMATCH,
                     });
@@ -212,16 +233,13 @@ export const infixOperationOperandsMustHaveCorrectType = (services: SafeDsServic
                 if (
                     node.rightOperand &&
                     !typeChecker.isSubtypeOf(rightType, coreTypes.Float) &&
-                    !typeChecker.isSubtypeOf(rightType, coreTypes.Int)
+                    !typeChecker.isSubtypeOf(rightType, coreTypes.Int) &&
+                    !typeChecker.isSubtypeOf(rightType, cellType)
                 ) {
-                    accept(
-                        'error',
-                        `Expected type '${coreTypes.Float}' or '${coreTypes.Int}' but got '${rightType}'.`,
-                        {
-                            node: node.rightOperand,
-                            code: CODE_TYPE_MISMATCH,
-                        },
-                    );
+                    accept('error', `This operator is not defined for type '${rightType}'.`, {
+                        node: node.rightOperand,
+                        code: CODE_TYPE_MISMATCH,
+                    });
                 }
                 return;
         }
@@ -338,10 +356,15 @@ export const prefixOperationOperandMustHaveCorrectType = (services: SafeDsServic
 
     return (node: SdsPrefixOperation, accept: ValidationAcceptor): void => {
         const operandType = typeComputer.computeType(node.operand);
+        const cellType = coreTypes.Cell(coreTypes.AnyOrNull);
+
         switch (node.operator) {
             case 'not':
-                if (!typeChecker.isSubtypeOf(operandType, coreTypes.Boolean)) {
-                    accept('error', `Expected type '${coreTypes.Boolean}' but got '${operandType}'.`, {
+                if (
+                    !typeChecker.isSubtypeOf(operandType, coreTypes.Boolean) &&
+                    !typeChecker.isSubtypeOf(operandType, cellType)
+                ) {
+                    accept('error', `This operator is not defined for type '${operandType}'.`, {
                         node,
                         property: 'operand',
                         code: CODE_TYPE_MISMATCH,
@@ -351,17 +374,14 @@ export const prefixOperationOperandMustHaveCorrectType = (services: SafeDsServic
             case '-':
                 if (
                     !typeChecker.isSubtypeOf(operandType, coreTypes.Float) &&
-                    !typeChecker.isSubtypeOf(operandType, coreTypes.Int)
+                    !typeChecker.isSubtypeOf(operandType, coreTypes.Int) &&
+                    !typeChecker.isSubtypeOf(operandType, cellType)
                 ) {
-                    accept(
-                        'error',
-                        `Expected type '${coreTypes.Float}' or '${coreTypes.Int}' but got '${operandType}'.`,
-                        {
-                            node,
-                            property: 'operand',
-                            code: CODE_TYPE_MISMATCH,
-                        },
-                    );
+                    accept('error', `This operator is not defined for type '${operandType}'.`, {
+                        node,
+                        property: 'operand',
+                        code: CODE_TYPE_MISMATCH,
+                    });
                 }
                 return;
         }